refactor: EventService interface 제거
